rest/requirements.txt
Flask~=2.0.3
gunicorn==20.1.0
flask-restx
google-cloud-storage
google-cloud-bigquery
Werkzeug~=2.0.3